repair: Move repair_row to a separate header.
